code

+++ src/main/java/com/data/respository/MovieDataRepo.java
@@ -4,7 +4,7 @@ package com.data.respository;
 import com.data.entity.MovieListData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieDataRepo extends JpaRepository<MovieListData,Long>
+public interface MovieDataRepo extends JpaRepository<MovieListData,Integer>
 {
 
 }
